
\name{as.capacity-methods}
\docType{methods}
\alias{as.capacity}
\alias{as.capacity-methods}
\alias{as.capacity,set.func-method}
\alias{as.capacity,card.capacity-method}

\title{Coercion method}

\description{Creates \bold{if possible} a new object of the desired class from the old object.}

\section{Methods}{
\describe{

\item{object = "set.func" }{Returns an object of class \code{capacity}.}

\item{object = "card.capacity" }{Returns an object of class \code{capacity}.}
}}

\keyword{methods}
%\keyword{ ~~ other possible keyword(s)}
