\name{as.card.capacity-methods}
\docType{methods}
\alias{as.card.capacity}
\alias{as.card.capacity-methods}
\alias{as.card.capacity,card.set.func-method}
\alias{as.card.capacity,capacity-method}

\title{Coercion method}

\description{Creates \bold{if possible} a new object of the desired class from the old object.}

\section{Methods}{
\describe{

\item{object = "card.set.func" }{Returns an object of class \code{card.capacity}.}

\item{object = "capacity" }{Returns an object of class \code{card.capacity}.}
}}

\keyword{methods}
%\keyword{ ~~ other possible keyword(s)}
