\name{as.card.game-methods}
\docType{methods}
\alias{as.card.game}
\alias{as.card.game-methods}
\alias{as.card.game,card.set.func-method}
\alias{as.card.game,game-method}

\title{Coercion method}

\description{Creates \bold{if possible} a new object of the desired class from the old object.}

\section{Methods}{
\describe{

\item{object = "card.set.func" }{Returns an object of class \code{card.game}.}

\item{object = "game" }{Returns an object of class \code{card.game}.}
}}

\keyword{methods}
%\keyword{ ~~ other possible keyword(s)}
