\name{as.game-methods}
\docType{methods}
\alias{as.game}
\alias{as.game-methods}
\alias{as.game,set.func-method}
\alias{as.game,card.game-method}

\title{Coercion method}

\description{Creates \bold{if possible} a new object of the desired class from the old object.}

\section{Methods}{
\describe{

\item{object = "set.func" }{Returns an object of class \code{game}.}

\item{object = "card.game" }{Returns an object of class \code{game}.}
}}
\keyword{methods}
%\keyword{ ~~ other possible keyword(s)}
