\name{as.Mobius.game-methods}
\docType{methods}
\alias{as.Mobius.game}
\alias{as.Mobius.game-methods}
\alias{as.Mobius.game,Mobius.set.func-method}

\title{Coercion method}

\description{Creates \bold{if possible} a new object of the desired class from the old object.}

\section{Methods}{
\describe{

\item{object = "Mobius.set.func" }{Returns an object of class \code{Mobius.game}.}
}}

\keyword{methods}
%\keyword{ ~~ other possible keyword(s)}
