\name{as.Mobius.set.func-methods}
\docType{methods}
\alias{as.Mobius.set.func}
\alias{as.Mobius.set.func-methods}
\alias{as.Mobius.set.func,Mobius.card.set.func-method}
\alias{as.Mobius.set.func,set.func-method}
\alias{as.Mobius.set.func,card.set.func-method}

\title{Coercion method}

\description{Creates \bold{if possible} a new object of the desired class from the old object.}

\section{Methods}{
\describe{

  \item{object = "Mobius.card.set.func" }{Returns an object of class
    \code{Mobius.set.func}.}

  \item{object = "set.func" }{Returns an object of class
    \code{Mobius.set.func}.}

  \item{object = "card.set.func" }{Returns an object of class
    \code{Mobius.set.func}.}
}}

\keyword{methods}
%\keyword{ ~~ other possible keyword(s)}
