\name{summary-methods}
\docType{methods}
\alias{summary}
\alias{summary-methods}
%\alias{summary,ANY-method}
\alias{summary,superclass.capacity-method}
\alias{summary,superclass.set.func-method}

\title{Summary method}

\description{Computes indices that can be used to summarize a set function.}

\section{Methods}{
\describe{

%\item{object = "ANY" }{}

\item{object = "superclass.capacity" }{Returns an object of class \code{summary.superclass.capacity}.}

\item{object = "superclass.set.func" }{Returns an object of class \code{summary.superclass.set.func}.}
}}

\seealso{
  	\code{\link{summary.superclass.capacity-class}}, 
	\cr \code{\link{summary.superclass.set.func-class}}. 
}

\keyword{methods}
%\keyword{ ~~ other possible keyword(s)}
